requirements.txt
pdfplumber>=0.6.0
scikit-learn>=1.2.0
pyyaml>=6.0
jsonschema>=4.0.0
click>=8.0.0
pytest>=7.0.0
sentence-transformers>=2.2.2
pytz>=2023.3